Remove useless mock class from main library
