+++ src/clabe/ui/__init__.py
@@ -1,4 +1,4 @@
-from .picker import DefaultPicker, PickerBase
+from .picker import PickerBase
 from .ui_helper import DefaultUIHelper, UiHelper, prompt_field_from_input
 
-__all__ = ["DefaultPicker", "PickerBase", "DefaultUIHelper", "UiHelper", "prompt_field_from_input"]
+__all__ = ["PickerBase", "DefaultUIHelper", "UiHelper", "prompt_field_from_input"]
