+++ src/clabe/ui/picker.py
@@ -243,55 +243,3 @@ class PickerBase(abc.ABC, Generic[_L, _R, _S, _T]):
         Subclasses should implement this method to perform any necessary cleanup operations.
         """
         ...
-
-
-class DefaultPicker(PickerBase[_L, _R, _S, _T]):
-    """
-    Default implementation of the picker. This serves as a placeholder implementation.
-
-    This class provides a basic implementation that raises NotImplementedError for
-    all picker methods, serving as a template for actual picker implementations.
-    """
-
-    def pick_rig(self) -> _R:
-        """
-        Raises NotImplementedError as this method is not implemented.
-
-        Raises:
-            NotImplementedError: Always, as this is a placeholder implementation
-        """
-        raise NotImplementedError("pick_rig method is not implemented")
-
-    def pick_session(self) -> _S:
-        """
-        Raises NotImplementedError as this method is not implemented.
-
-        Raises:
-            NotImplementedError: Always, as this is a placeholder implementation
-        """
-        raise NotImplementedError("pick_session method is not implemented")
-
-    def pick_task_logic(self) -> _T:
-        """
-        Raises NotImplementedError as this method is not implemented.
-
-        Raises:
-            NotImplementedError: Always, as this is a placeholder implementation
-        """
-        raise NotImplementedError("pick_task_logic method is not implemented")
-
-    def initialize(self) -> None:
-        """
-        Placeholder implementation for initialization.
-
-        Does nothing in the default implementation.
-        """
-        return
-
-    def finalize(self) -> None:
-        """
-        Placeholder implementation for finalization.
-
-        Does nothing in the default implementation.
-        """
-        return
